requirements.txt
aio-pika==9.3.0
aioredis==2.0.1
aiormq==6.7.7
async-timeout==4.0.3
idna==3.4
multidict==6.0.4
pamqp==3.2.1
pika==1.3.2
python-dotenv==1.0.0
typing_extensions==4.8.0
yarl==1.9.2
